requirements.txt
flask==3.0.3
line-bot-sdk==3.11.0
anthropic==0.40.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
